grub: Update grub git recipe
Update the grub git recipe with the following actions for native and target

- Base class extend to native

This is required because on aarch64 we need native recipe to be compiled for host architecture and generate grub-mkimage, which is
executed on host machine to generate grubaa64.efi binary for target architecture. In the case of x86/x86_64 architecture
it does not make any difference because host=target. The flow on aarch64 is as below

grub-native compile for host (generates grub-mkimage) ==> do_install_append_class-target() generates grubaa64.efi binary by executing host executable binary grub-mkimage ==> do_deploy() deploy the grubaa64.efi

Hence add the below lines

+DEPENDS_class-target += "grub-native"
+BBCLASSEXTEND = "native"

- Add autotools-brokensep

The recipe is fails to work (because B!=S) when I have add this recipe to luvOS, hence inherit autotools-brokensep instead of autotools.
May be this needs to be fixed.

- On x86 architecture the recipe builds only for PC, It may be required to extend this recipe.

Signed-off-by: Naresh Bhat <naresh.bhat@linaro.org>